Extract Java interface
