+++ mubench.pipeline/data/java_interface.py
@@ -1,0 +1,34 @@
+from enum import Enum, IntEnum
+from logging import Logger
+from typing import Optional, List, Dict
+
+from data.project_version import ProjectVersion
+from utils.shell import Shell
+
+def _quote(value: str):
+    return "\"{}\"".format(value)
+
+class JavaInterface:
+    def __init__(self, jar_path: str, java_options: List[str]):
+        self.jar_path = jar_path
+        self.java_options = java_options
+
+    @staticmethod
+    def get(cli_version: str, jar_path: str, java_options: List[str]) -> 'JavaInterface':
+        return JavaInterfaceV0(jar_path, java_options)
+
+    def execute(self, version: ProjectVersion, detector_arguments: Dict[str, str],
+                timeout: Optional[int], logger: Logger):
+        raise NotImplementedError
+
+class JavaInterfaceV0(JavaInterface):
+    def execute(self, version: ProjectVersion, detector_arguments: Dict[str, str],
+                timeout: Optional[int], logger: Logger):
+        command = self._get_command(version, detector_arguments)
+        Shell.exec(command, logger=logger, timeout=timeout)
+
+    def _get_command(self, version: ProjectVersion, detector_arguments: Dict[str, str]) -> str:
+        detector_invocation = ["java"] + self.java_options + ["-jar", _quote(self.jar_path)]
+        command = detector_invocation + detector_arguments
+        command = " ".join(command)
+        return command
